Refactored Kafka consumer
